[Chore] #60 - 리뷰 반영 struct enum 파일 분리 및 preview 코드 삭제
